// src/node/config/config-type.ts
export enum TextMode {
  PLAIN_TEXT = "text/plain",
  SQL = "text/sql",
  JAVASCRIPT = "text/javascript",
  PYTHON = "python",
  SHELL = "shell",
  JSON = "application/json",
}

export enum ConfigurationType {
  STRING = "STRING",
  TAGS = "TAGS",
  NUMBER = "NUMBER",
  BOOLEAN = "BOOLEAN",
  TEXT_AREA = "TEXT_AREA",
  SELECT = "SELECT",
  LABEL = "LABEL",
  KNOWLEDGE_BASE = "KNOWLEDGE_BASE",
}

export enum PanelType {
  CONVENTIONAL = "CONVENTIONAL",
  NODE_CONFIG = "NODE_CONFIG",
  RUN_OPTION = "RUN_OPTION",
}